move TorServiceConnection to it's own file
